SIteheld and aliquots model update
