+++ summit_blood_samples/blood_sample/templates/home.html
@@ -114,7 +114,7 @@ $(document).ready(function(){
                             var chart = this,
                             series = this.series[0],
                             xAxis = chart.xAxis[0],
-                            newStart = series.xData[ (series.xData.length-12) ],
+                            newStart = series.xData[ (series.xData.length-14) ],
                             newEnd = series.xData[ (series.xData.length) ];
                             xAxis.setExtremes(newStart, newEnd);
                         }
@@ -268,7 +268,7 @@ $(document).ready(function(){
                             var chart = this,
                             series = this.series[0],
                             xAxis = chart.xAxis[0],
-                            newStart = series.xData[ (series.xData.length-12) ],
+                            newStart = series.xData[ (series.xData.length-14) ],
                             newEnd = series.xData[ (series.xData.length) ];
                             xAxis.setExtremes(newStart, newEnd);
                         }
@@ -393,7 +393,7 @@ $(document).ready(function(){
                             var chart = this,
                             series = this.series[0],
                             xAxis = chart.xAxis[0],
-                            newStart = series.xData[ (series.xData.length-12) ],
+                            newStart = series.xData[ (series.xData.length-14) ],
                             newEnd = series.xData[ (series.xData.length) ];
                             xAxis.setExtremes(newStart, newEnd);
                         }
@@ -542,7 +542,7 @@ $(document).ready(function(){
                             var chart = this,
                             series = this.series[0],
                             xAxis = chart.xAxis[0],
-                            newStart = series.xData[ (series.xData.length-12) ],
+                            newStart = series.xData[ (series.xData.length-14) ],
                             newEnd = series.xData[ (series.xData.length) ];
                             xAxis.setExtremes(newStart, newEnd);
                         }
