+++ summit_blood_samples/blood_sample/templates/review_blood_sample/blood-sample-review.html
@@ -19,11 +19,6 @@
         <div id="pagination_blood_sample" class="pagination-md pull-right"></div></br>
     </div>
     <em class="error"></em>
-    {% comment %} {% if navigation != True %}
-        <div class="text-right bottom-buttons">
-            <button type="button" class="btn btn-cancel" data-dismiss="modal">Cancel</button>
-        </div>
-    {% endif %} {% endcomment %}
 </div>
 <div class="site-content__main custom-modal-content edit-sample-div">
 </div>
