+++ summit_blood_samples/blood_sample/templates/review_processed/pr_unmatched-table.html
@@ -8,6 +8,7 @@
             <th class="processed-header">Processed Date Time</th>
             <th class="processed-header">Volume</th>
             <th class="processed-header">Number Of Children</th>
+            <th class="processed-header">Site Held</th>
             <th class="processed-header">Comments</th>
             <th>Edit</th>
         </tr>
@@ -18,6 +19,7 @@
             <td class="processed-col">{{ item.ProcessedDateTime }}</td>
             <td class="processed-col">{{ item.Volume }}</td>
             <td class="processed-col">{{ item.NumberOfChildren }}</td>
+            <td class="processed-col">{{ item.SiteHeld }}</td>
             <td class="processed-col"><a rel="tooltip" data-html="true" title="<h6>{{ item.Comments }}</h6>" >{{ item.Comments|truncatechars:20 }}</a></td>
             <td><a data-editprocessed="{{item.id}}" class="edit-unmatched-processed" style="cursor:pointer;"><img class="warning-svg" src="{% static 'indigo/images/edit.svg' %}" style="margin-bottom: 0px;width:20px;height: 20px;" /></a></td>
         </tr>
